// types.ts
export enum UserRole {
  GROWER = 'CULTIVADOR',
  ADMIN = 'ADMINISTRADOR',
  MAINTENANCE = 'MANTENIMIENTO',
  TRIMMER = 'TRIMEADOR',
}
// ...
export enum PlantBatchStatus {
    GERMINATION_ROOTING = 'Germinación/Enraizamiento',
    PRE_VEGETATION = 'Pre-vegetación',
    IN_USE = 'En Cultivo',
    DEPLETED = 'Agotado',
}

export interface PlantBatch {
  id: string; 
  name: string; 
  geneticsId: string;
  creationDate: string;
  initialPlantCount: number;
  rootedPlantCount?: number; // Renamed from finalPlantCount for clarity
  availablePlantCount: number; // For tracking available plants for new crops
  sourceLocationId: string; // e.g., 'room-ss1', 'mother-plant-1'
  type: 'seed' | 'clone';
  status: PlantBatchStatus;
  creatorId: string; // User ID
  logEntries: LogEntry[];
}

export interface MotherPlant {
  id: string;
  name: string; // e.g., "OG Kush Madre #1"
  geneticsId: string;
  locationId: string; // Room ID where it is located
  sowingDate: string; // ISO date string
  cloneCount: number; // Total clones taken from this plant
  logEntries: LogEntry[];
  isArchived?: boolean;
}

export interface SensorDataPoint {
    timestamp: string;
    temperature: number;
    humidity: number;
}

export interface LogEntry {
  id: string;
  date: string; // ISO Date string
  environmental?: {
    temp: number;
    humidity: number;
    vpd: number;
    co2: number;
    leafTemp?: number;
  };
  irrigation?: {
    type: 'nutrients' | 'supplements';
    ph: number;
    ppm: number;
    volume: number;
    phOut?: number;
    ppmOut?: number;
    cost?: number;
  };
  foliarSpray?: { name: string; dose: string }[];
  supplements?: { name: string; dose: string }[];
  notes?: string;
  plantHealth?: string[];
  photos?: string[]; // base64 string array
  completedTasks?: {
    taskId: string;
    taskTitle: string;
    completedBy: string; // User Username
    completedAt: string; // ISO Date string
  }[];
  sensorData?: SensorDataPoint[];
}